Update: Modularise Tag component
